+++ CHANGELOG.md
@@ -4,6 +4,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [unreleased]
 
+- Feature: Add `PillNumberInput` component.
 - Feature: Add `micro` size for `<Button icon>`.
 - Feature: Add original `file <File>` on `onChange` callback for `<ImageCropper />`